requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiohttp>=3.8.5
asyncio>=3.4.3
python-dotenv>=1.0.0 
anthropic>=0.15.0  # Claude API client

starlette==0.46.0
typing-extensions==4.13.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
requests
# mcp-sdk==0.1.0  # Uncomment when the SDK is available 
mcp[cli] >= 0.4.0  